split up create_templates script into a lib file
