+++ lib/static/views.js
@@ -232,11 +232,11 @@ exports.setup = function(app) {
 
   // /common/js/templates.js is dynamically built each time
   if (!config.get('use_minified_resources')) {
-    var generateTemplates = require('../../scripts/create_templates');
+    var templates = require('../templates');
     var dialogTemplatesPath = path.join(__dirname, '../../resources/static/dialog/views')
     var dialogTemplatesData;
     app.get('/common/js/templates.js', function(req, res) {
-      var str = generateTemplates(generateTemplates.RETURN, dialogTemplatesPath);
+      var str = templates.generate(dialogTemplatesPath);
       if (str) dialogTemplatesData = str;
 
       res.send(dialogTemplatesData);
@@ -245,7 +245,7 @@ exports.setup = function(app) {
     var siteTemplatesPath = path.join(__dirname, "../../resources/views");
     var siteTemplatesData;
     app.get('/test/mocks/site-templates.js', function(req, res) {
-      var str = generateTemplates(generateTemplates.RETURN, siteTemplatesPath, "site/");
+      var str = templates.generate(siteTemplatesPath, "site/");
       if (str) siteTemplatesData = str;
 
       res.send(siteTemplatesData);
